debt - move outputAppender to where it is used
