+++ src/tsconfig.strictNullChecks.json
@@ -324,7 +324,6 @@
 		"./vs/platform/notification/common/notification.ts",
 		"./vs/platform/notification/test/common/testNotificationService.ts",
 		"./vs/platform/opener/common/opener.ts",
-		"./vs/platform/output/node/outputAppender.ts",
 		"./vs/platform/progress/common/progress.ts",
 		"./vs/platform/quickOpen/common/quickOpen.ts",
 		"./vs/platform/quickinput/common/quickInput.ts",
@@ -573,6 +572,7 @@
 		"./vs/workbench/contrib/output/common/output.ts",
 		"./vs/workbench/contrib/output/common/outputLinkComputer.ts",
 		"./vs/workbench/contrib/output/common/outputLinkProvider.ts",
+		"./vs/workbench/contrib/output/node/outputAppender.ts",
 		"./vs/workbench/contrib/performance/electron-browser/startupTimings.ts",
 		"./vs/workbench/contrib/preferences/browser/settingsWidgets.ts",
 		"./vs/workbench/contrib/preferences/common/smartSnippetInserter.ts",
