+++ src/vs/workbench/contrib/output/node/outputAppender.ts
